Initial packaging of dnf-plugins-core
Summary:
This is the packaging for the core plugins for DNF,
the next upcoming major version of Yum.

The core plugins include many useful commands for enhancing
the base functionality of the DNF package manager.

It is required for updating Mock, as the builddep subcommand
is implemented as a plugin that is included in this package.

Reviewers: ikey

Reviewed By: ikey

Subscribers: kyrios123, ikey

Differential Revision: https://dev.solus-project.com/D2116